projects/users/pyproject.toml
[tool.poetry]
name = "users"
version = "0.1.0"
description = ""
authors = ["Brayan Henao <b.henaoc@uniandes.edu.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
boto3 = "^1.34.74"
sse-starlette = "^2.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = "^2.6.4"
bcrypt = "^4.1.2"


[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
faker = "^24.4.0"
coverage = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
